Rework file list to allow objects to be returned.
The file list in the native system was a list of the leaf names,
but in some systems (eg Pyromaniac) we can return a list of objects
containing all the files that were present with their attributes.
We now support this as it's more efficient.

SVG handling is better as well - we don't assume the files is in
the CWD.