Make the function naming for FSExplorerFrame more WxPython-y.
Trying to make the naming for the functions used by the FSExplorerFrame
more like the way that the other WxPython functions work. There's going
to be a number of things I get wrong but I think it's better than it was.
Maybe I should use Events more than methods on the object, but it'll do
for now.