Allow FSExplorers objects to directly open windows.
In order to make it easier to get the same semantics from the window
opening in the existing windows, as the initial window open, the
FSExplorers object now has a `open_window` function to allow it to
either open an existing window or open a new one. This still has to
happen on the UI thread, but it's quite a sane interface I think.

Plus it means that I can, in the future, make the file info opening
work.