Add support for reading timestamp of files; better info box layout.
We can now read the timestamp of the files and present this in the
information box.

The file info box layout has been exceptionally hard. It shouldn't
be but reading the size of the list when it was the UltimateListCtrl
was found to be impossible, and even with the ListCtrl it's being
harder than you might expect.