Add support for the Full Info display mode.
The way that these displayed icons are managed is not at all efficient.
Although the new Full info display mode has been created, it isn't fast
and that means that it's not really useful with large windows. However
we're only trying to make it possible to find things, not to create a
perfect filer replacement.