Make the sort options propagate to new windows.
When a new explorer window is opened from an existing window, it
propagates the sort format into the new window. This keeps it more
like the Filer, although the original Filer just had a single global
setting that it changed when you selected the format and applied to
new windows.