Rudimentary support for creating directories.
We now have rudimentary support in the explorer for the new directory
menu option, which will call down to the FS interfaces to try to create
a directory. We have functions to check whether the directory can be
created too.

The handling hasn't been properly tested yet, and the fsnative
implementation doesn't have support.