Add error handler for if directory creation fails.
When the directory creation fails we need to report this to the
user, so we now have a method to report errors in a useful way -
this could be overridden by changing the function, to make other
styles of error box. In Pyromaniac I may use this to report the
error in a RISC OS-style box (although that might be confusing
as it's not an error /from/ RISC OS).