Fixes from RISC OS Pyromaniac; addition of Rename option.
Numerous fixes from testing within RISC OS Pyromaniac. The menus now work
better, and error reports don't crash. The text in the windows is now
black and we've got better reports for failing to create directories.

There's a new Rename option and we have the option of support, like the
other menu and FS operations.