Improve the file information window to not obscure as easily.
The file information window would obscure the text of what it was
showing when the window was resized, with the scrollbars. This would
not reset when the window was sized larger. This appered to be more
obvious on the Windows version, as it seemed to be forced from being
opened. There is now a little more space and the code uses a sizer
around the panel as well, so that it's more in line with common
use.