Fix for failure to run on Python 3.
The code should cope with running on Python 2 and Python 3, however, it
wasn't even trying. The checks for str or unicode will be able to work
now - we generate the unicode type, just to allow that instance check
to work. The wx.Size calls weren't working when they were being passed
floats rather than ints. We now force the values passed to it to be
ints.